refactor: move and renamew useIsCurrent
